+++ src/js/components/slider.ts
@@ -1,0 +1,13 @@
+import Swiper from 'swiper';
+
+function init() {
+
+	var Slider = function (container: string, options: any) {
+		return new Swiper(container, options);
+	};
+
+	(<any>window).Slider = Slider;
+}
+
+
+export default { init };