+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const CopyPlugin = require('copy-webpack-plugin');
 
-const UglifyJsPlugin = require("uglifyjs-webpack-plugin");
+const TerserPlugin = require('terser-webpack-plugin');
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
 
@@ -90,8 +90,7 @@ module.exports = {
 
 	optimization: {
 		minimize: true,
-		minimizer: [
-			new UglifyJsPlugin({
+		minimizer: [new TerserPlugin({
 				cache: true,
 				parallel: true,
 				sourceMap: true // set to true if you want JS source maps